+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `BlockTreeEntry::ancestor` to look up an ancestor block at a given height. Returns `None` if the height is out of range. This operation is O(log N).
+- Added `Transaction::locktime()` to retrieve a transaction's `nLockTime` value as a `u32`.
 
 ## [0.2.0] 2026-01-26